Extract author determination code so it's easier to test
